requirements.txt
matplotlib==3.3.2
Pillow==8.1.0
mysql-connector==2.2.9
requests==2.22
pandas==1.2.1